Un-did unnecessary file renames
